Auto register entities in DB as networked
Periodically attempt to register non-networked entities in the DB as
networked.

There is a limit on the number of networked objects that are synced, and
when this limit is reached, new objects spawned will not be networked.
This ensures that, as room becomes available, those objects will be
sychnronized with other players.